+++ src/QuickMarcEditor/QuickMarcEditorRows/LinkButton/LinkButton.js
@@ -1,12 +1,12 @@
 import {
   useState,
   useMemo,
+  useContext,
 } from 'react';
 import {
   useIntl,
   FormattedMessage,
 } from 'react-intl';
-import { useLocation } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import uniq from 'lodash/uniq';
 
@@ -26,6 +26,7 @@ import {
 } from '@folio/stripes/components';
 import { useAuthorityLinkingRules } from '@folio/stripes-marc-components';
 
+import { QuickMarcContext } from '../../../contexts';
 import { useMarcSource } from '../../../queries';
 import { MarcFieldContent } from '../../../common';
 import {
@@ -68,12 +69,11 @@ const LinkButton = ({
 }) => {
   const stripes = useStripes();
   const intl = useIntl();
-  const location = useLocation();
   const [authority, setAuthority] = useState(null);
   const [isModalOpen, setIsModalOpen] = useState(false);
   const callout = useCallout();
+  const { isSharedRef } = useContext(QuickMarcContext);
   const centralTenantId = stripes.user.user?.consortium?.centralTenantId;
-  const isSharedBibRecord = new URLSearchParams(location.search).get('shared') === 'true';
 
   let showSharedFilter = false;
   let showSharedRecordsOnly = false;
@@ -84,7 +84,7 @@ const LinkButton = ({
       showSharedRecordsOnly = true;
     }
   } else if (checkIfUserInMemberTenant(stripes)) {
-    if (isSharedBibRecord) {
+    if (isSharedRef.current) {
       if (action === QUICK_MARC_ACTIONS.EDIT) {
         showSharedRecordsOnly = true;
         pluginTenantId = centralTenantId;