+++ src/QuickMarcEditor/QuickMarcEditorRows/QuickMarcEditorRows.js
@@ -96,7 +96,6 @@ const QuickMarcEditorRows = ({
   onCheckCentralTenantPerm = noop,
   onInputFocus,
 }) => {
-  const location = useLocation();
   const stripes = useStripes();
   const intl = useIntl();
   const { initialValues } = useFormState();
@@ -105,7 +104,10 @@ const QuickMarcEditorRows = ({
   const newRowRef = useRef(null);
   const rowContentWidth = useRef(null); // for max-width of resizable textareas
   const childCalloutRef = useRef(null);
-  const { validationErrorsRef } = useContext(QuickMarcContext);
+  const {
+    validationErrorsRef,
+    isSharedRef,
+  } = useContext(QuickMarcContext);
 
   const {
     linkAuthority,
@@ -114,7 +116,7 @@ const QuickMarcEditorRows = ({
     autoLinkableBibFields,
   } = useAuthorityLinking({ marcType, action });
 
-  const isRequestToCentralTenantFromMember = applyCentralTenantInHeaders(location, stripes, marcType)
+  const isRequestToCentralTenantFromMember = applyCentralTenantInHeaders(isSharedRef.current, stripes, marcType)
     && action === QUICK_MARC_ACTIONS.EDIT;
 
   const fixedFieldInitialValues = () => {