+++ src/QuickMarcEditor/useSaveRecord/useSubmitRecord/useSumbitRecord.js
@@ -60,6 +60,7 @@ const useSubmitRecord = ({
     instance,
     continueAfterSave,
     relatedRecordVersion,
+    setIsShared,
   } = useContext(QuickMarcContext);
 
   const { actualizeLinks } = useAuthorityLinking({ marcType, action });
@@ -87,7 +88,11 @@ const useSubmitRecord = ({
     // when a user creates a new Bib or Authority in a central tenant - it becomes shared
     // so we need to append this parameter to the URL to tell quickMARC it is now a shared record
     if (isInCentralTenant && marcType !== MARC_TYPES.HOLDINGS) {
-      searchParams.append('shared', true);
+      setIsShared(true);
+    }
+
+    if (action === QUICK_MARC_ACTIONS.DERIVE) {
+      setIsShared(false);
     }
 
     const routes = {
@@ -96,13 +101,13 @@ const useSubmitRecord = ({
       [MARC_TYPES.HOLDINGS]: `${basePath}/edit-holdings/${externalId}`,
     };
 
-    await refreshPageData(fieldIds, QUICK_MARC_ACTIONS.EDIT, externalId);
-
-    history.push({
+    await history.push({
       pathname: routes[marcType],
       search: searchParams.toString(),
     });
-  }, [basePath, marcType, location, history, refreshPageData, stripes]);
+
+    await refreshPageData(fieldIds, QUICK_MARC_ACTIONS.EDIT, externalId);
+  }, [basePath, marcType, location, history, refreshPageData, stripes, action, setIsShared]);
 
   const onCreate = useCallback(async (formValues, _api) => {
     const formValuesToProcess = prepareForSubmit(formValues);