+++ src/hooks/useAuthorityLinking/useAuthorityLinking.js
@@ -1,5 +1,6 @@
 import {
   useCallback,
+  useContext,
   useMemo,
 } from 'react';
 import { useLocation } from 'react-router-dom';
@@ -30,15 +31,16 @@ import {
   UNCONTROLLED_NUMBER,
   QUICK_MARC_ACTIONS,
 } from '../../QuickMarcEditor/constants';
+import { QuickMarcContext } from '../../contexts';
 
 const formatSubfieldCode = (code) => { return code.startsWith('$') ? code : `$${code}`; };
 
 const useAuthorityLinking = ({ tenantId, marcType, action } = {}) => {
   const stripes = useStripes();
-  const location = useLocation();
+  const { isSharedRef } = useContext(QuickMarcContext);
 
   const centralTenantId = stripes.user.user.consortium?.centralTenantId;
-  const isCentralTenantInHeaders = applyCentralTenantInHeaders(location, stripes, marcType)
+  const isCentralTenantInHeaders = applyCentralTenantInHeaders(isSharedRef.current, stripes, marcType)
     && action === QUICK_MARC_ACTIONS.EDIT;
 
   // tenantId for linking functionality must be with the member tenant id when user derives shared record